Fix width list in font descriptor
The widths were stored in a HashMap, which means they won't be in order
when looping over them. This changes them to a Vec since no lookup by
gid is needed.

Issue #40 was caused because the insertion of the widths to `width_list`
is delayed by one iteration and the last remaining widths weren't
inserted. That combined with the arbitrary order of the HashMap
iteration causes the broken widths do occur randomly.

Resolves: #40