Allow users to reduce optimization
For many use cases pdfs are write once read many, but in my specific
case pdfs are written once and read roughly once or twice. I would
prefer to disable compression and other optimizations in some cases.

In addition to not compressing when debug assertions are enabled, some
compressions are disabled when the feature "less-optimization" is
enabled.