Remove deps.rs badge from README.md
Unfortunately it seems that deps.rs is not maintained anymore (https://github.com/srijs/deps.rs/issues/40), and it has failed to keep up-to-date with the latest crates.io API, which has prevented it from analysing dependencies. It is potentially worth removing the badge until such a time that the leadership of the deps.rs project is clarified and changes are deployed.