Use WinAnsiEncoding for built-in fonts
Previously, PdfLayerReference::write_text would use the UTF-8 encoding
for built-in fonts.  But the built-in fonts only support a very limited
set of encodings, see Appendix D of the PDF Reference [0].  When adding
the font to the PDF document, printpdf already selects the
WinAnsiEncoding (Windows-1252, see the Into<LoDictionary> implementation
for BuiltinFont).  This would lead to encoding issues for all non-ASCII
characters, see also issue #30.

With this patch, we use lopdf::Document::encode_text to encode the text
for built-in fonts with the WinAnsiEncoding.  This fixes the encoding
issues for built-in fonts.  Note that that function just drops
characters that are not supported by the encoding.

This patch also adds notes about this behaviour to the documentation of
the add_builtin_font, use_text and write_text methods.

Fixes #30.

[0] https://www.adobe.com/content/dam/acom/en/devnet/pdf/pdfs/pdf_reference_archives/PDFReference.pdf