Use BTreeMap in FontList for reproducibility
This patch replaces the HashMap in FontList with a BTreeMap so that the
fonts in the generated PDF document are added in sorted by their name.
This makes it easier to reproducibly generate PDF files e. g. in unit
tests.

To be able to use IndirectFontRef as the key of a BTreeMap, we also have
to derive Ord.