Use OffsetDateTime::now_utc instead of now
OffsetDateTime::now has been deprecated in time 0.2.11 since it does not
make it clear that the UTC time zone is used.  (Still, for printpdf, UTC
is used intentionally, see the to_pdf_time_stamp_metadata function in
document_info.rs.)   With this patch, we use now_utc instead of now.  As
now_utc was added in time 0.2.11, we have to update the minimum
dependency version.